requirements.txt
Django>=3.1.7
djangorestframework>=3.12.4
psycopg2-binary>=2.8.6
openmeteo-requests>=1.2.0
requests-cache>=1.2.0
pandas
retry-requests>=2.0.0
django-db-views>=0.1.6